microservices/auth/requirements.txt
flask
pyjwt 